+++ pages/genres/[...slug].js
@@ -7,7 +7,7 @@ const Slug = () => {
     <div>
       <div className="bg-gradient-to-tr from-rose-50 to-rose-200 dark:bg-gradientedge text-gray-900 dark:text-gray-100 min-h-screen">
         <Header />
-        <UnsupportedRoute />
+        <UnsupportedRoute url={`https://www.goodreads.com/genres/${slug}`} />
         <Footer />
       </div>
     </div>